display_rawfb: rotate mouse input
In case monitor is rotated, the input coordinates from the
mouse must be rotated, too. The temporary buffer where the
pointer is displayed must be computed according to the rotated
coordinates instead of just dumping from the framebuffer.

Showing the pointer requires to store and restore the background
of the pointer, but this should also taken into account the rotation
of the coordinates.

Signed-off-by: Stefano Babic <sbabic@denx.de>